+++ src/components/LoginForm/LoginForm.module.css
@@ -1,30 +1,67 @@
-/* .form {
+.form {
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 14px;
-  padding: 0 24px 24px 24px;
+  /* padding: 0 24px 0 24px; */
 }
+
 .button {
-  color: 161616;
+  font-size: 14px;
+  font-weight: 500;
+  margin-top: 10px;
+  line-height: 21px;
+  letter-spacing: -0.02em;
+  color: #161616;
   background-color: #bedbb0;
   border-radius: 8px;
   width: 287px;
   height: 49px;
   border: none;
   cursor: pointer;
+  transition: background-color 400ms ease-in-out;
 }
+
+
 .button:hover {
   background-color: #9dc888;
 }
+ 
+.inputContainer {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
 .input {
   background-color: #151515;
-  border: 1px solid #bedbb0;
+  border: 1px solid rgba(255, 255, 255, 0.4);
   border-radius: 8px;
   width: 287px;
   height: 49px;
-  padding: 10px;
+  padding: 14px 40px 14px 18px;
   margin-bottom: 15px;
   font-size: 16px;
-  outline: #1F1F1F;
-} */
+  outline: none;
+  margin: 0;
+  color:rgba(255, 255, 255, 0.4);
+}
+
+.input::placeholder{
+  color: rgba(255, 255, 255, 0.4);
+}
+
+.icon {
+  position: absolute;
+  right: 18px;
+  top: 50%;
+  transform: translateY(-50%);
+  cursor: pointer;
+  color:  rgba(255, 255, 255, 0.4); 
+}
+
+.errorMessage {
+  color: #BEDBB0;
+  padding: 5px 0 5px 18px;
+}
