+++ src/components/RegisterForm/RegisterForm.module.css
@@ -1,9 +1,9 @@
-/* .form {
+.form {
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 14px;
-  padding: 0 24px 0 24px;
+  /* padding: 0 24px 0 24px; */
 }
 
 .button {
@@ -15,6 +15,7 @@
   height: 49px;
   border: none;
   cursor: pointer;
+  transition: background-color 400ms ease-in-out;
 }
 
 .button:hover {
@@ -53,4 +54,9 @@
   transform: translateY(-50%);
   cursor: pointer;
   color:  rgba(255, 255, 255, 0.4); 
-} */
+}
+
+.errorMessage {
+  color: #BEDBB0;
+  padding: 5px 0 5px 18px;
+}
