+++ src/components/Welcome/Welcome.jsx
@@ -6,9 +6,8 @@ import clsx from "clsx";
 
 const Welcome = () => {
   return (
-    <div className={css.background}>
       <div className={css.section}>
-      <div className={css.container}>
+      <div className={css.fix}>
         <div className={css.avatar}>
           <img
             className={css.developerAvatar}
@@ -21,10 +20,11 @@ const Welcome = () => {
           <Icon id="icon-main-logo" width="40" height="40" className={css.iconApp} />
           <h1 className={css.title}>Task Pro</h1>
         </div>
-        <p className={css.description}>
-          Supercharge your productivity and take control of your tasks with Task
-          Pro - Don&apost wait, start achieving your goals now!
+        <p className={clsx(css.description, css.multiLine)}>
+          Supercharge your productivity and take control <br className={css.breakLine} /> of your tasks with Task
+          Pro - Don&apos;t wait, start achieving your goals now!
         </p>
+        <div className={css.btnBox}>
         <Link to="/auth/register" className={css.authLink}>
           <button type="submit" className={clsx(css.button, css.buttonActive)}>
             Registration
@@ -35,8 +35,8 @@ const Welcome = () => {
             Log in
           </button>
         </Link>
+        </div>
       </div>
-    </div>
     </div>
   );
 };
