+++ src/components/Welcome/Welcome.module.css
@@ -1,8 +1,3 @@
-.background {
-  background: linear-gradient(to bottom, #ffffff, #bedbb0);
-  width: 100vh;
-}
-
 .section {
   margin: 0;
   align-items: center;
@@ -11,19 +6,20 @@
   flex-direction: column;
   flex-wrap: wrap;
   height: 100vh;
+  background: linear-gradient(to bottom, #ffffff, #bedbb0);
 }
-.container {
-  min-width: 320px;
-  max-width: 375px;
-  margin-left: auto;
-  margin-right: auto;
-  padding-left: 20px;
-  padding-right: 20px;
+
+@media only screen and (min-width: 768ps) {
+  .fix {
+    width: 473px;
+  }
 }
+
 .avatar {
   display: flex;
   justify-content: center;
 }
+
 .logoName {
   display: flex;
   justify-content: center;
@@ -49,12 +45,19 @@
 }
 .description {
   font-weight: 400;
+  word-wrap: break-word; 
   font-size: 14px;
-  line-height: 129%;
+  line-height: 18px;
   letter-spacing: -0.02em;
   text-align: center;
   color: #161616;
   padding: 0 0 48px 0;
+}
+
+.btnBox {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .button {
@@ -72,15 +75,12 @@
   transition: background-color 400ms ease-in-out;
 }
 
-/* .button:hover,
-.button:focus {
-  background-color: #161616; 
-  color: #ffffff;
-}
+@media only screen and (min-width: 768px){
+  .button {
+  width: 344px;
+}}
 
-.buttonActive {
-  background-color: #9DC888; 
-} */
+
 
 .buttonActive {
   background-color: #161616; 
@@ -97,3 +97,14 @@
   background-color: #9DC888; 
   color: #ffffff;
 }
+
+.multiLine {
+  white-space: pre-line;
+}
+
+@media (max-width: 375px) {
+  .breakLine {
+    display: none;
+  }
+}
+
